AAdvance 100

+++ Assets/ProjectAssets/Scripts/SO/PlayerScoreData.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerScoreData", menuName = "Scriptable Objects/PlayerScoreData")]
 public class PlayerScoreData : ScriptableObject
 {
-    public float Score;
+    public int Score;
 }
